added class ExportUtils.py, increased lens blur
