+++ Application/VQAInteractionScreen.py
@@ -21,6 +21,8 @@ from docx.shared import Inches, Pt
 from worker import Worker
 from utils import PredictionResults, predictVilt, predictLxmert
 
+from ExportUtils import ExportUtils
+
 class VQAInteractionScreen(QWidget):
     def __init__(self, threadManager, controller, models, parent=None):
         super().__init__(parent)
@@ -28,13 +30,7 @@ class VQAInteractionScreen(QWidget):
         self.controller = controller
         self.models = models
         self.currentImage = None
-        self.currentQuestion = ""
-        self.currentModelImage = None
         self.predictionResult = None
-        self.exported_dir = "Exports and Snapshots/"
-        self.snapshot_dir = "Exports and Snapshots/Snapshots/"
-        self.model_dir = "Exports and Snapshots/Model Results/"
-        self.current_doc_name = ""
         self.current_model_details = ""
         self.visuals = []
         self.cameraEffect = "None"
@@ -47,7 +43,7 @@ class VQAInteractionScreen(QWidget):
                                 ["Maple Leaves", 0.0], ["Road Leaves", 0.0], 
                                 ["Dust", 0.0], ["Fog", 0.0]
                             ]
-        self.model_used = ""
+        self.ExportResults = ExportUtils()
         self.load_ui()
       
     def load_ui(self):
@@ -101,7 +97,7 @@ class VQAInteractionScreen(QWidget):
         self.ui.pushButton_FreezeUnfreezeFrame.clicked.connect(self.freezeUnfreezeCamera)
 
         # Take a Snapshot Button Clicked
-        self.ui.pushButton_TakeASnapshot.clicked.connect(self.takeSnapshot)
+        self.ui.pushButton_TakeASnapshot.clicked.connect(lambda: self.ExportResults.takeSnapshot(self.ui.pushButton_TakeASnapshot, self.currentImage))
 
         # Ask Question Button
         self.ui.pushButton_Ask.clicked.connect(self.askQuestion)
@@ -111,7 +107,7 @@ class VQAInteractionScreen(QWidget):
         valLabel.setText(str(sliderVal))
 
         self.weatherEffects[command][1] = sliderVal
-        
+
         # Convert slider value from [0-100] to [0.00-1.00] range equivalent & pass to AirSim controller
         decSliderVal = sliderVal / 100
         self.controller.updateAirSimWeather(command, decSliderVal)
@@ -147,7 +143,7 @@ class VQAInteractionScreen(QWidget):
             self.cameraEffect = "Black Screen"
 
         if (self.ui.radioButton_LensBlur.isChecked()):
-            frame = cv2.GaussianBlur(frame, (5,5), 10.0)
+            frame = cv2.GaussianBlur(frame, (15,15), cv2.BORDER_DEFAULT)
             self.cameraEffect = "Lens Blur"
 
         if (self.ui.radioButton_PixelCorruption.isChecked()):
@@ -175,53 +171,6 @@ class VQAInteractionScreen(QWidget):
             self.timer.start(100)
             self.ui.pushButton_FreezeUnfreezeFrame.setText("Freeze Frame")
     
-    def getCurrentFormattedTime(self): 
-        current_time = time.localtime()
-        formatted_time = time.strftime("%b-%d-%H-%M-%S", current_time)
-        return formatted_time
-
-    def checkExportDir(self):
-        try:
-            if (not os.path.exists(self.exported_dir)):
-                os.mkdir(self.exported_dir)
-            
-            if (not os.path.exists(self.snapshot_dir)):
-                os.mkdir(self.snapshot_dir)
-            
-            if (not os.path.exists(self.model_dir)):
-                os.mkdir(self.model_dir)
-            
-            return True
-        except:
-            return False
-
-
-    def takeSnapshot(self):
-        self.ui.pushButton_TakeASnapshot.setText("Capturing")
-        self.ui.pushButton_TakeASnapshot.setEnabled(False)
-
-        formatted_time = self.getCurrentFormattedTime()
-        filename = formatted_time + ".png"
-        
-        if (not self.checkExportDir()):
-            print("Could not save snapshot")
-            return
-        
-        try: 
-            cv2.imwrite(self.snapshot_dir + filename, cv2.cvtColor(self.currentImage, cv2.COLOR_RGB2BGR))
-            print("Snapshot saved to: " + self.snapshot_dir + filename)
-
-            self.ui.pushButton_TakeASnapshot.setText("Captured")
-            Timer(2, self.resetTakeSnapshot).start()
-        except:
-            print("Could not save snapshot")
-        
-    
-    def resetTakeSnapshot(self):
-        self.ui.pushButton_TakeASnapshot.setText("Take a Snapshot")
-        self.ui.pushButton_TakeASnapshot.setEnabled(True)
-
-
     def askQuestion(self):
         # Hide any existing displayed visualization
         self.ui.label_ResultVisualization.hide()
@@ -233,22 +182,17 @@ class VQAInteractionScreen(QWidget):
 
         # Obtain the desired model for prediction
         question = self.ui.lineEdit_Question.text()
-        self.currentQuestion = question
         image = self.currentImage.copy()
-        self.currentModelImage = image
 
         model_index = 0
-        self.model_used = ""
         # ViLT (Base) Model
         if self.ui.radioButton_ViltBase.isChecked():
             model_index = 0
-            self.model_used = "ViLT"
             model = self.models[model_index]
             worker = Worker(predictVilt, model[0], model[1], question, image) 
         # LXMERT (Base) Model
         elif self.ui.radioButton_LxmertBase.isChecked():
             model_index = 1
-            self.model_used = "LXMERT"
             model = self.models[model_index]
             worker = Worker(predictLxmert, model[0], model[1], model[2], model[3], model[4], question, image)
 
@@ -341,104 +285,6 @@ class VQAInteractionScreen(QWidget):
             self.displayVisualization(defaultImageIndex)
         
         if (self.ui.checkBox_ExportResults.isChecked()):
-            self.exportResults()
-
-    def resetExportText(self):
-        self.ui.checkBox_ExportResults.setText("Export Results")
-
-    def exportResults(self):
-
-            if (not self.checkExportDir()):
-                print("Could Not Export")
-                return
-
-            try: 
-                document = Document()
-                formatted_time = self.getCurrentFormattedTime()
-
-                result_dir = self.model_dir + formatted_time + "/"
-
-                os.mkdir(result_dir)
-
-                self.current_doc_name = self.model_dir + formatted_time + "/Results_" + formatted_time + ".docx"
-
-                document.add_heading("Question and Model Prediction")
-
-                answer_paragraph = document.add_paragraph()
-                answer_format = answer_paragraph.paragraph_format
-                answer_format.left_indent = Inches(0.5)
-
-                modelUsed_run = answer_paragraph.add_run("Model Used: " + self.model_used)
-                modelUsed_font = modelUsed_run.font
-                modelUsed_font.name = 'Calibri'
-                modelUsed_font.size = Pt(12)
-
-                question_run = answer_paragraph.add_run("\nQuestion Asked: " + self.currentQuestion)
-                question_font = question_run.font
-                question_font.name = 'Calibri'
-                question_font.size = Pt(12)
-
-                answer_run = answer_paragraph.add_run("\nPrediction: " + self.predictionResult.prediction)
-                answer_font = answer_run.font
-                answer_font.name = 'Calibri'
-                answer_font.size = Pt(12)
-
-                if (self.current_model_details != ""):
-                    document.add_heading("Prediction Details")
-                    details_paragraph = document.add_paragraph()
-                    details_format = details_paragraph.paragraph_format
-                    details_format.left_indent = Inches(0.5)
-
-                    details_run = details_paragraph.add_run(self.current_model_details)
-                    details_font = details_run.font
-                    details_font.name = 'Calibri'
-                    details_font.size = Pt(12)
-
-                document.add_heading("User Settings")
-                settings_paragraph = document.add_paragraph()
-                settings_format = settings_paragraph.paragraph_format
-                settings_format.left_indent = Inches(0.5)
-
-                cameraEffect_run = settings_paragraph.add_run("Camera Effect: " + self.cameraEffect)
-                cameraEffect_font = cameraEffect_run.font
-                cameraEffect_font.name = 'Calibri'
-                cameraEffect_font.size = Pt(12)
-
-                for weather_condition in self.weatherEffects: 
-                    print(weather_condition[0], weather_condition[1])
-                    if (weather_condition[1] > 0.0):
-                        weather_run = settings_paragraph.add_run("\n" + weather_condition[0] + ": " + str(weather_condition[1]) + "%")
-                        weather_font = weather_run.font
-                        weather_font.name = 'Calibri'
-                        weather_font.size = Pt(12)
-
-                document.add_heading("Base Image")
-                cv2.imwrite(result_dir + "base_image.png", cv2.cvtColor(self.currentModelImage, cv2.COLOR_RGB2BGR))
-                document.add_picture(result_dir + "base_image.png", width=Inches(6.0))
-
-                numVisuals = len(self.predictionResult.visualizations)
-                if numVisuals:
-                    document.add_heading("Model Visualizations")
-                    # Export each visual
-                    for i in range(numVisuals):
-                        # Resize Image for Export (uses size from GUI)
-                        dim = (self.ui.label_ResultVisualization.width(),self.ui.label_ResultVisualization.height())
-                        export_frame = cv2.resize(self.predictionResult.visualizations[i], dim)
-
-                        visual_filename = f"Visualization_{i+1}.png"
-                        cv2.imwrite(result_dir + visual_filename, cv2.cvtColor(export_frame, cv2.COLOR_RGB2BGR))
-                        
-                        document.add_heading(f"Visualization_{i+1}.png", level=2)
-                        document.add_picture(result_dir + visual_filename, width=Inches(6.0))
-                
-
-                document.save(self.current_doc_name)
-                print("results exported to: " + self.current_doc_name)
-
-                self.ui.checkBox_ExportResults.setText("Exported")
-                Timer(2, self.resetExportText).start()
-            except:
-                print("Could not export results")
-                self.ui.checkBox_ExportResults.setText("Could Not Export")
-                Timer(2, self.resetExportText).start()
-
+            self.ExportResults.exportResults(self.predictionResult, self.current_model_details, 
+                                             self.cameraEffect, self.weatherEffects, 
+                                             self.ui.checkBox_ExportResults)
